+++ src/main/java/org/openlmis/fulfillment/i18n/MessageKeys.java
@@ -83,6 +83,7 @@ public abstract class MessageKeys {
   public static final String PERMISSION_MISSING = join(ERROR_PREFIX, PERMISSION, MISSING);
   public static final String PERMISSIONS_MISSING = join(ERROR_PREFIX, PERMISSIONS, MISSING);
 
+  public static final String ORDER_EXISTS = join(ERROR_PREFIX, ORDER, EXISTS);
   public static final String ORDER_NOT_FOUND = join(ERROR_PREFIX, ORDER, NOT_FOUND);
   public static final String ORDER_INVALID_STATUS = join(ERROR_PREFIX, ORDER, INVALID_STATUS);